pyproject.toml
[tool.poetry]
name = "e621"
version = "0.0.2"
description = "e621.net API wrapper written in Python"
authors = ["PatriotRossii <patriotrossii2019@mail.ru>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/PatriotRossii/e621-py"
keywords = ["e621", "furry", "yiff", "api"]
classifiers = ["Operating System :: OS Independent"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.27.1"
pydantic = "^1.9.0"
"backports.cached-property" = "^1.0.1"
typing-extensions = "^4.1.1"

[tool.poetry.dev-dependencies]
datamodel-code-generator = "^0.11.20"
black = "^22.3.0"
isort = "^5.10.1"
typer = "^0.4.1"
pytest = "^7.1.1"
pytest-mock = "^3.7.0"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
